kuvar/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
clearbit==0.1.7
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
Markdown==3.3.3
pyhunter==1.7
PyJWT==2.0.0
pytz==2020.5
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.2
